Add factorial function

+++ FormulaParser/DivideFunctionDescriptor.cpp
@@ -4,7 +4,7 @@
 
 double DivideFunctionDescriptor::Evaluate(const vector<unique_ptr<TreeItem>>& items) const {
 	double result = UserFunctions::Evaluate(items[0]);
-    for (int i = 1; i < items.size(); i++) {
+    for (size_t i = 1; i < items.size(); i++) {
         auto res = UserFunctions::Evaluate(items[i]);
         if (res == 0)
             throw EvaluateFormulaException("Division by zero");
